# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Список",
                "operationId": "47925884fb256655799d0ffd57c2a894",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Some title"
                                                    },
                                                    "likes": {
                                                        "type": "integer",
                                                        "example": 200
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Создание",
                "operationId": "7999e1ba2e5585f522fc5d9d27915557",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 200
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "likes": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts/{post}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Еденичная запись",
                "operationId": "c5253d16752aa5889a1fa6e763471311",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "likes": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Удаление",
                "operationId": "6089ca174ecd6f787461d01e729e93c8",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Post"
                ],
                "summary": "Обновление",
                "operationId": "58e7e5c3e0c1e1f02bc488b3be5eb91b",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID поста",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title for edit"
                                            },
                                            "likes": {
                                                "type": "integer",
                                                "example": 21
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Some title"
                                                },
                                                "likes": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Post",
            "description": "Post"
        }
    ]
}